third_party/blink/web_tests/platform/generic/editing/pasteboard/paste-into-table-cell-2-expected.txt

This tests pasting into a table cell.
To manually run the test, copy some text and paste it after "1".
The content should be pasted into the first cell, not the second cell.
| <table>
|   border="1"
|   <tbody>
|     <tr>
|       <td>
|         id="firstCell"
|         "1 2<#selection-caret>"
|       <td>
|         "3"
